% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtualPollen.R
\docType{data}
\name{accumulationRate}
\alias{accumulationRate}
\title{Accumulation rate}
\format{dataframe with 10000 rows and the following columns:
#' @return A dataframe with the following columns.
\itemize{
  \item \emph{time} numeric, time or age of the given case.
  \item \emph{accumulation.rate} numeric, in years per centimetre, simulated accumulation rate.
  \item \emph{grouping} integer, grouping variable to aggregate together (with \code{\link{aggregateSimulation}}) samples deposited in the same centimetre according \emph{accumulation.rate}.
}}
\usage{
data(accumulationRate)
}
\description{
Dataframe, output of \code{\link{simulateAccumulationRate}}
}
\seealso{
\code{\link{simulateAccumulationRate}}, \code{\link{aggregateSimulation}}
}
\author{
Blas M. Benito  <blasbenito@gmail.com>
}
\keyword{datasets}
