% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtualPollen.R
\docType{data}
\name{driverA}
\alias{driverA}
\title{Driver A}
\format{Numeric vector of length 5000.}
\usage{
data(driverA)
}
\description{
A vector of 5000 values (years) between 0 and 100 generated by \code{\link{simulateDriver}} with a temporal autocorrelation significant for 600 years. It is meant to be used as input for \code{\link{simulatePopulation}}.
}
\seealso{
\code{\link{simulateDriver}}
}
\author{
Blas M. Benito  <blasbenito@gmail.com>
}
\keyword{datasets}
