% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtualPollen.R
\docType{data}
\name{driverB}
\alias{driverB}
\title{Driver B}
\format{numeric vector of length 10000.}
\usage{
data(driverB)
}
\description{
A vector of 10000 values (years) between 0 and 100 generated by \code{\link{simulateDriver}} with a temporal autocorrelation significant for 600 years. It is meant to be used as input for \code{\link{simulatePopulation}}.
}
\seealso{
\code{\link{simulateDriver}}
}
\author{
Blas M. Benito  <blasbenito@gmail.com>
}
\keyword{datasets}
