% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtualPollen.R
\docType{data}
\name{drivers}
\alias{drivers}
\title{Drivers with different temporal autocorrelations.}
\format{numeric vector of length 10000.}
\usage{
data(drivers)
}
\description{
A dataframe with 60000 rows and 4 columns (years) containing two drivers (A and B) generated by \code{\link{simulateDriver}} with different temporal autocorrelations (200, 600, and 1800). It is meant to be used as input for \code{\link{simulatePopulation}}
}
\details{
\itemize{
  \item \emph{time} integer.
  \item \emph{driver} character, values are \code{A} and \code{B}
  \item \emph{autocorrelation.length} numeric, values are 200, 600, and 1800.
  \item \emph{value} numeric, value of the driver for the given \emph{time}.
}
}
\seealso{
\code{\link{simulateDriver}}
}
\author{
Blas M. Benito  <blasbenito@gmail.com>
}
\keyword{datasets}
