% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametersCheck.R
\name{parametersCheck}
\alias{parametersCheck}
\title{Plots main simulation parameters.}
\usage{
parametersCheck(parameters, species="all", driver.A=NULL, driver.B=NULL, drivers=NULL, filename=NULL)
}
\arguments{
\item{parameters}{the parameters dataframe.}

\item{species}{if "all" or "ALL", all species in "parameters" are plotted. It also accepts a vector of numbers representing the rows of the selected species, or a vector of names of the selected species.}

\item{driver.A}{numeric vector with driver values.}

\item{driver.B}{numeric vector with driver values.}

\item{drivers}{dataframe with drivers}

\item{filename}{character string, filename of the output pdf.}
}
\value{
A \code{\link{ggplot2}} object.
}
\description{
Plots the normal function/s, fecundity, growth curve, and maturity age, of each virtual taxa in \code{parameters}.
}
\details{
The function prints the plot, can save it to a pdf file if \code{filename} is provided, and returns a \code{\link{ggplot2}} object.
}
\examples{
#generating driver
driver <- simulateDriver(random.seed = 10, time = 1:1000, autocorrelation.length = 200, output.min = 0, output.max = 100, rescale = TRUE)

#preparing parameters
parameters <- parametersDataframe(rows=2)
parameters[1,] <- c("Species 1", 50, 20, 2, 0.2, 0, 100, 10000, 1, 0, 50, 10, 0, 0, 600, 0)
parameters[2,] <- c("Species 2", 500, 100, 10, 0.02, 0, 100, 10000, 1, 0, 50, 10, 0, 0, 600, 0)
parameters <- fixParametersTypes(x=parameters)

#plotting parameters
parametersCheck(parameters=parameters, driver.A=driver, filename="Parameters.pdf")

}
\seealso{
\code{\link{parametersDataframe}}, \code{\link{fixParametersTypes}}
}
\author{
Blas M. Benito  <blasbenito@gmail.com>
}
