% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAcf.R
\name{plotAcf}
\alias{plotAcf}
\title{Plots results of \code{\link{acfToDf}}.}
\usage{
plotAcf(
  x = NULL,
  plot.title = ""
  )
}
\arguments{
\item{x}{dataframe, output of \code{\link{acfToDf}}}

\item{plot.title}{string, title for the plot.}
}
\value{
A ggplot object
}
\description{
Plots a dataframe resulting from \code{\link{acfToDf}} by using \code{\link{ggplot2}} (and \code{\link{cowplot}})
}
\examples{

#simulating driver
x <- simulateDriver(
  random.seed = 10,
  time = 1:1000,
  autocorrelation.length = 200,
  output.min = -10,
  output.max = 20,
  rescale = TRUE
  )

#computing temporal autocorrelation
x.df <- acfToDf(
  x = x,
  lag.max = 300,
  length.out = 100
  )

#plotting output
plotAcf(x.df)

}
\seealso{
\code{\link{acfToDf}}
}
\author{
Blas M. Benito  <blasbenito@gmail.com>
}
