% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acfToDf.R
\name{acfToDf}
\alias{acfToDf}
\title{Applies \code{\link{acf}} to a vector and returns a dataframe for pretty plotting in \code{\link[ggplot2]{ggplot2}}.}
\usage{
acfToDf(
  x = NULL,
  lag.max = 100,
  length.out = 10
  )
}
\arguments{
\item{x}{numeric vector.}

\item{lag.max}{integer, number of lags over which to compute temporal autocorrelation.}

\item{length.out}{integer, total number of lags to consider for plotting. Should be a subset of \code{lag.max}.}
}
\value{
A dataframe with the columns:
#' \itemize{
  \item \code{lag}: numeric, lag in the time units of \code{x} with a maximum determined by \code{lag.max}, and a number of unique values determined by \code{length.out}
  \item \code{acf}: Pearson correlation index returned by the \code{\link{acf}} for a given number of lags for the given lag.
  \item \code{ci.max}: Maximum value of the confidence interval of \code{acf}.
  \item \code{ci.min}: Minimum value of the confidence interval of \code{acf}.
}
}
\description{
It reads a vector representing a time series, applies \code{\link{acf}} for a given number of lags
}
\details{
This function computes temporal autocorrelation in a given vector using \code{\link{acf}}, and returns a dataframe ready for easy plotting in \code{\link{ggplot2}}.
}
\examples{

#simulating driver
x <- simulateDriver(
  random.seed = 10,
  time = 1:1000,
  autocorrelation.length = 200,
  output.min = -10,
  output.max = 20,
  rescale = TRUE
  )

#computing temporal autocorrelations
x.df <- acfToDf(
  x = x,
  lag.max = 300,
  length.out = 100
  )
str(x.df)

#plotting output
plotAcf(x.df)

}
\seealso{
\code{\link{acf}}
}
\author{
Blas M. Benito  <blasbenito@gmail.com>
}
