% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSimulation.R
\name{plotSimulation}
\alias{plotSimulation}
\title{Plots results of \code{\link{simulatePopulation}}.}
\usage{
plotSimulation(
  simulation.output=NULL,
  species="all",
  burnin=FALSE,
  filename=NULL,
  time.zoom=NULL,
  panels=c("Driver A",
           "Driver B",
           "Suitability",
           "Population",
           "Mortality",
           "Biomass",
           "Pollen"
           ),
 plot.title=NULL,
 width=12,
 text.size=20,
 title.size=25,
 line.size=1
 )
}
\arguments{
\item{simulation.output}{output of \code{\link{simulatePopulation}}.}

\item{species}{a number or vector or numbers representing rows in the parameters dataframe, or a string or vector of strings referencing to the "label" column of the parameters dataframe.}

\item{burnin}{if \code{FALSE}, burn-in period is not considered in the model.}

\item{filename}{character string, name of output pdf file. If NULL or empty, no pdf is produced. It shouldn't include the extension of the output file.}

\item{time.zoom}{vector of two numbers indicating the beginnign and end of the time interval to be plotted (i.e. "c(5000, 10000)")}

\item{panels}{character string or vector of character strings with these possible values: "Driver A", "Driver B","Suitability", "Population", "Mortality", "Biomass", "Pollen".}

\item{plot.title}{character string to use as plot title.}

\item{width}{plot width in inches.}

\item{text.size}{text size of the plot.}

\item{title.size}{plot title size.}

\item{line.size}{size of lines in plots.}
}
\description{
This function takes as input a list of dataframes or a single dataframe resulting from the execution of \code{\link{simulatePopulation}}, and plots the resulting time series of pollen abundance, number of individuals, biomass, driver, and environmnetal suitability.
}
\details{
The user can decide what virtual taxa to plot (argument \code{species}), and what information to show throught the \code{panels} argument. Output is plotted on screen by default, and printed to pdf if the \code{filename} argument is filled.
}
\examples{

driver <- simulateDriver(
 random.seed = 10,
 time = 1:1000,
 autocorrelation.length = 200,
 output.min = 0,
 output.max = 100,
 rescale = TRUE
 )

#preparing parameters
parameters <- parametersDataframe(rows = 2)
parameters[1,] <- c("Species 1", 50, 20, 2, 0.2, 0, 100, 1000, 1, 0, 50, 10, 0, 0, NA, NA)
parameters[2,] <- c("Species 1", 500, 100, 10, 0.02, 0, 100, 1000, 1, 0, 50, 10, 0, 0, NA, NA)
parameters <- fixParametersTypes(x = parameters)

#simulating population dynamics
sim.output <- simulatePopulation(
 parameters = parameters,
 driver.A = driver
 )

#plot simulation
plotSimulation(simulation.output = sim.output)

}
\seealso{
\code{\link{simulatePopulation}}, \code{\link{compareSimulations}}
}
\author{
Blas M. Benito  <blasbenito@gmail.com>
}
